+++ soya/src/compiler/webpack/WebpackCompiler.js
@@ -208,10 +208,10 @@ export default class WebpackCompiler extends Compiler {
           //  'imports': ['react'],
           //  'locals':  ['module']
           //},
-          {
-            'transform': 'react-transform-catch-errors',
-            'imports': ['react', 'redbox-react']
-          }
+          //{
+          //  'transform': 'react-transform-catch-errors',
+          //  'imports': ['react', 'redbox-react']
+          //}
         ]
       };
     }