+++ test/integration/src/pages/redux/RuntimeComponent/RuntimeComponent.js
@@ -9,7 +9,7 @@ import UserSegment from '../../../segments/user/UserSegment.js';
 import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 
 // TODO: Figure out how to do promise polyfill.
-import { Promise } from 'es6-promise';
+////import { Promise } from 'es6-promise';
 
 import style from '../../../shared/sitewide.css';