+++ test/integration/src/segments/LifetimeSessionSegment.js
@@ -6,21 +6,20 @@ import Thunk from 'soya/lib/data/redux/Thunk.js';
 //import { Promise } from 'es6-promise';
 import request from 'superagent';
 
-export default class UserSegment extends MapSegment {
+export default class LifetimeSessionSegment extends MapSegment {
   static id() {
     return 'user';
   }
 
   _generateQueryId(query) {
-    return query.username;
+    return 'default';
   }
 
   _generateThunkFunction(thunk) {
-    var query = thunk.query;
     var queryId = thunk.queryId;
     thunk.func = (dispatch) => {
       var result = new Promise((resolve, reject) => {
-        request.get('http://localhost:8000/api/user/' + query.username).end((err, res) => {
+        request.get('http://localhost:8000/api/context').end((err, res) => {
           if (res.ok) {
             var payload = JSON.parse(res.text);
             dispatch(this._createSyncLoadActionObject(queryId, payload));